Update README.md
I spoke with Armindo last week. He helped me resolve 2 issues in the read.me. One typo, and one area of clarification to the OBS text string. Being unfamiliar with the structure, I left in the brackets.

'portraits" is a typo. The main.py is looking instead for "portrait"
remove the extension from the image name; i.e. my-image.png should be my-image
remove all brackets; my working text string is http://127.0.0.1:5000/portrait?name=raine&id=123456789 replace the id with your own
make sure the server is running
If that's not working, I suggest looking at the error in the server.
A 400 likely means that it's either not receiving the character name or the character id.
A 404 is exactly that; file not found.
Although Amindo had to help with the bracket thing, using the errors and playing with the script before hand got us there pretty quickly.